pyproject.toml
[project]
name = "telecom-customer-churn"
version = "0.1.0"
description = "End-to-end MLOps solution for predicting customer churn in telecommunications, combining traditional ML with AI-driven text enrichment."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Sebastian" }]

dependencies = [
    # --- Core ML ---
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.5",
    "optuna>=4.0",
    "xgboost>=2.1",
    "lightgbm>=4.5",
    "imbalanced-learn>=0.12",
    # --- Experiment Tracking ---
    "mlflow>=2.18",
    # --- Data Validation ---
    "great-expectations>=1.3",
    # --- NLP / Embeddings ---
    "sentence-transformers>=3.3",
    # --- AI Agent / LLM ---
    "google-genai>=1.0",
    "langchain>=0.3",
    "langchain-google-genai>=2.0",
    "langgraph>=0.2",
    "pydantic-ai>=0.0.32",
    # --- API / Serving ---
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    # --- UI ---
    "gradio>=5.0",
    # --- Utilities ---
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "python-box>=7.2",
    "pydantic>=2.10",
    "rich>=13.9",
    "httpx>=0.28",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "logfire>=4.28.0",
    "pytest-asyncio>=0.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "pyright>=1.1",
    "ipykernel>=6.29",
    "notebook>=7.3",
]
dvc = [
    "dvc>=3.56",
    "dvc-s3>=3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["notebooks/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "W", "PTH"]
ignore = ["E501", "W293"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pyright]
include = ["src"]
pythonVersion = "3.11"
typeCheckingMode = "standard"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=src --cov-report=term-missing"
asyncio_mode = "auto"

